refractor(relational): abstract the common method
