+++ dataSource/relational/src/main/resources/dameng-meta.properties
@@ -100,6 +100,10 @@ jdbc_support_special_char=false
 
 support_boundary_query=true
 
+avg_cast_expression_base=%s(CAST(%s AS DECIMAL(34, 16)))
+
+avg_cast_expression_default=%s(%s)
+
 # 单表最大列数
 max_column_num_limit=2000
