+++ dataSource/relational/src/main/resources/mysql-meta.properties
@@ -81,6 +81,14 @@ max_single_row_size_limit=7900
 
 support_boundary_query=true
 
+avg_cast_expression_base=%s(CAST(%s AS DECIMAL(34, 16)))
+
+avg_cast_expression_default=%s(%s)
+
+boundary_query_catalog_column=table_schema
+boundary_query_schema_column=table_schema
+boundary_query_use_db_connection=false
+
 # 配置MySQL DataTypeTransformer
 TINYINT=IGinX-BOOLEAN
 SMALLINT=IGinX-INTEGER
