+++ dataSource/relational/src/main/resources/oceanbase-meta.properties
@@ -65,6 +65,14 @@ max_column_num_limit=800
 # 单行最大字节数
 max_single_row_size_limit=8195
 
+avg_cast_expression_base=%s(CAST(%s AS DECIMAL(34, 16)))
+avg_cast_expression_default=%s(%s)
+
+support_native_upsert=true
+
+boundary_query_catalog_column=table_schema
+boundary_query_schema_column=table_schema
+
 # 配置MySQL DataTypeTransformer
 TINYINT=IGinX-BOOLEAN
 SMALLINT=IGinX-INTEGER
