+++ dataSource/relational/src/main/resources/oracle-meta.properties
@@ -92,6 +92,10 @@ jdbc_support_get_table_name_from_result_set=false
 
 support_boundary_query=true
 
+avg_cast_expression_base=%s(CAST(%s AS BINARY_DOUBLE))
+
+avg_cast_expression_default=%s(CAST(%s AS BINARY_DOUBLE))
+
 # 单表最大列数
 max_column_num_limit=950
