+++ dataSource/relational/src/main/resources/postgresql-meta.properties
@@ -64,6 +64,14 @@ jdbc_support_special_char=true
 schema_pattern=public
 # 是否支持 SQL2003 规定的 information_schema 视图
 support_boundary_query=true
+# baseExpression和其他Expression的avg函数表达式
+avg_cast_expression_base=%s(%s)
+avg_cast_expression_default=%s(%s)
+# 查询元数据边界时，是否需要针对不同的数据库建立独立的连接
+boundary_query_use_db_connection=true
+# 查询元数据边界的table_catalog字段名和table_schema字段名
+boundary_query_catalog_column=table_catalog
+boundary_query_schema_column=table_schema
 # 单表最大列数
 max_column_num_limit=1550
 # 单行最大字节数
